% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_predict.R
\name{binary_predict}
\alias{binary_predict}
\title{Binary prediction}
\usage{
binary_predict(PI)
}
\arguments{
\item{PI}{is the vector containing the probabilities of assignment to the positive or negative class.}
}
\value{
a vector containing the predictions in binary form
}
\description{
This function performs binary predictions (0 or 1) based on the probabilities of assignments.
}
\examples{
\dontrun{
   binary_predict(PI)
}
}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
