% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.reduction.R
\name{centering.red.pred}
\alias{centering.red.pred}
\title{centering and reduction of the test set}
\usage{
centering.red.pred(X, archive_moy_ect)
}
\arguments{
\item{X}{the test set dataframe to center and reduce}

\item{archive_moy_ect}{list containing the mean and standard deviation of each column of the training set}
}
\value{
the dataframe of the test set centered and reduced
}
\description{
centering and reduction of the test set
}
\examples{
arch <- data.frame(Sepal.Length=c(mean(iris[1:100,1]),
sqrt((nrow(iris)-1)/var(iris[1:100,1]))),Sepal.Width=c(mean(iris[1:100,2]),
sqrt((nrow(iris)-1)/var(iris[1:100,2]))))
centering.red.pred(iris[101:150,c(1,2)],arch)
}
