% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.reduction.R
\name{centering.reduction}
\alias{centering.reduction}
\title{centering and reduction}
\usage{
centering.reduction(X)
}
\arguments{
\item{X}{the dataframe to center and reduce}
}
\value{
the dataframe centered and reduce
}
\description{
centering and reduction
}
\examples{
centering.reduction(iris[,-ncol(iris)])
}
