% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupage_colonne.R
\name{decoupage_colonne}
\alias{decoupage_colonne}
\title{Column Decoupage}
\usage{
decoupage_colonne(data, ncores)
}
\arguments{
\item{data}{dataframe to cut by column}

\item{ncores}{Cutting in ncores parts}
}
\value{
list of data frame
}
\description{
Column Decoupage
}
\examples{
\dontrun{
 decoupage_colonne(data, ncores)
}
}
