% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupage_ligne.R
\name{decoupage_ligne}
\alias{decoupage_ligne}
\title{Algorithm to cut data}
\usage{
decoupage_ligne(data, ncores)
}
\arguments{
\item{data}{data to cut}

\item{ncores}{split data ncores parts}
}
\value{
a dataframe with an additional column indicating the group of each observation
}
\description{
Algorithm to cut data
}
\examples{
\dontrun{
 decoupage_ligne(data,ncores)
}
}
