% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_minibatch_online_seq.R
\name{dg_batch_minibatch_online_seq}
\alias{dg_batch_minibatch_online_seq}
\title{Global Gradient descent algorithm}
\usage{
dg_batch_minibatch_online_seq(
  X,
  y,
  theta,
  batch_size,
  random_state,
  leaning_rate,
  max_iter,
  tolerance,
  rho = NA,
  C = NA
)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict.}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{batch_size}{a parameter that specifies the number of observations in each mini-batch. It can significantly affect performance}

\item{random_state}{this parameter defines the seed of the random number generator, use when shuffling to mix observations.}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}

\item{rho}{hyper parameter which allows arbitration between RDIGE and LASSO.}

\item{C}{parameter allowing to arbitrate between the penalty and the likelihood in the guidance of the modeling.}
}
\value{
this function returns an instance containing:
\itemize{
\item final theta
\item history cost
\item iteration number
\item intern iteration number
}
}
\description{
This function allows the execution of the binary logistic regression according to the batch, mini batch and online mode.
}
\examples{
\dontrun{
 global_grad_descent(X,y,theta)
}
}
