% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_batch_seq.R
\name{dg_batch_seq}
\alias{dg_batch_seq}
\title{Gradient descent algorithm}
\usage{
dg_batch_seq(X, y, theta, leaning_rate, max_iter, tolerance, rho = NA, C = NA)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict.}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}

\item{rho}{hyper parameter which allows arbitration between RDIGE and LASSO.}

\item{C}{parameter allowing to arbitrate between the penalty and the likelihood in the guidance of the modeling.}
}
\value{
this function returns the instance of model with all parameters
}
\description{
Gradient descent algorithm
}
\examples{
\dontrun{
  gradient(X,y,theta)
  gradient(X,y,theta, leaning_rate=0.1, max_iter=100, tolerance=1e-04)
}
}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
