% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgrglm.fit.R
\name{dgrglm.fit}
\alias{dgrglm.fit}
\title{Function fit to construct model}
\usage{
dgrglm.fit(
  formule,
  data,
  ncores = NA,
  mode_compute = "parallel",
  leaning_rate = 0.1,
  max_iter = 100,
  tolerance = 1e-04,
  batch_size = NA,
  random_state = 102,
  centering = FALSE,
  feature_selection = FALSE,
  p_value = 0.01,
  rho = 0.1,
  C = 0.1,
  iselasticnet = FALSE
)
}
\arguments{
\item{formule}{allows you to define the target variable and predictor variables}

\item{data}{the data source containing all the variables specified in the formula}

\item{ncores}{parameters representing the number of cores to be used for parallel execution}

\item{mode_compute}{algorithm mode of execution. thera are "parallel" and "sequentiel"}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update}

\item{max_iter}{is the number of iterations}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}

\item{batch_size}{a parameter that specifies the number of observations in each mini-batch. It can significantly affect performance}

\item{random_state}{this parameter defines the seed of the random number generator, use when shuffling to mix observations}

\item{centering}{to center and reduce the variables, by default FALSE}

\item{feature_selection}{parameters indicating the choice to make the selection of variable or not}

\item{p_value}{selection criteria}

\item{rho}{hyper parameter which allows arbitration between RDIGE and LASSO. Elasticnet case.}

\item{C}{parameter allowing to arbitrate between the penalty and the likelihood in the guidance of the modeling.Elasticnet case.}

\item{iselasticnet}{for Elasticnet}
}
\value{
an instance of model
}
\description{
This function allows us to create a binary logistic regression model
}
\examples{
\dontrun{
 dgrglm.fit(formule, data)
 dgrglm.fit(formule, data,ncores=3, mode_compute="parallel")
}
}
