% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgrglm.multiclass.predict.R
\name{dgrglm.multiclass.predict}
\alias{dgrglm.multiclass.predict}
\title{Binary or probabilities prediction}
\usage{
dgrglm.multiclass.predict(model, new_data)
}
\arguments{
\item{model}{the model created with fit function logistic regression}

\item{new_data}{the new data to be classified}
}
\value{
an instance containing the binary or probability prediction.
}
\description{
Binary or probabilities prediction
}
\examples{
\dontrun{
 predict(model, new_data, type_pred)
}
}
