% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgrglm.predict.R
\name{dgrglm.predict}
\alias{dgrglm.predict}
\title{Binary or probabilities prediction}
\usage{
dgrglm.predict(model, new_data, type_pred = "CLASS", centering = FALSE)
}
\arguments{
\item{model}{the model created with fit function logistic regression}

\item{new_data}{the new data to be classified}

\item{type_pred}{the type of prediction to perform (binary prediction or probability prediction) by default 'CLASS'}

\item{centering}{to center and reduce the variables, by default FALSE}
}
\value{
an instance containing the binary or probability prediction.
}
\description{
Binary or probabilities prediction
}
\examples{
\dontrun{
 predict(model, new_data, type_pred, centering)
}
}
