% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgsrow_batch_minibatch_online_parallele.R
\name{dgs_minibatch_online_parallle}
\alias{dgs_minibatch_online_parallle}
\title{Batch Mini & Online DGSRow Distributed}
\usage{
dgs_minibatch_online_parallle(
  X,
  y,
  theta,
  ncores,
  batch_size,
  leaning_rate,
  max_iter,
  tolerance
)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{ncores}{parameters representing the number of cores to be used for parallel execution}

\item{batch_size}{a parameter that specifies the number of observations in each mini-batch. It can significantly affect performance}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}
}
\value{
this function returns an instance containing:
\itemize{
\item final theta
\item history cost
\item iteration number
}
}
\description{
Batch Mini & Online DGSRow Distributed
}
\examples{
\dontrun{
 dgs_minibatch_online_parallle(X,y,theta)
 dgs_minibatch_online_parallle(X,y,theta,ncores=3)
}
}
