% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgsrow_minibatch_parallele_bis.R
\name{dgsrow_minibatch_parallle2}
\alias{dgsrow_minibatch_parallle2}
\title{MiniBatch DGSRow Distributed}
\usage{
dgsrow_minibatch_parallle2(
  X,
  y,
  theta,
  batch_size,
  ncores,
  leaning_rate,
  max_iter,
  tolerance
)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{batch_size}{a parameter that specifies the number of observations in each mini-batch. It can significantly affect performance}

\item{ncores}{parameters representing the number of cores to be used for parallel execution}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}
}
\value{
this function returns an instance containing:
\itemize{
\item final theta
\item history cost
\item iteration number
\item iteration number minibatch
}
}
\description{
MiniBatch DGSRow Distributed
}
\examples{
\dontrun{
 dgsrow_minibatch_parallle_bis(X,y,theta)
 dgsrow_minibatch_parallle_bis(X,y,theta,ncores=3)
}
}
