% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Function fit to construct model}
\usage{
fit(
  formule,
  data,
  leaning_rate = 0.1,
  max_iter = 100,
  tolerance = 1e-04,
  batch_size = NA,
  random_state = 1
)
}
\arguments{
\item{formule}{allows you to define the target variable and predictor variables}

\item{data}{the data source containing all the variables specified in the formula}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}

\item{batch_size}{a parameter that specifies the number of observations in each mini-batch. It can significantly affect performance}

\item{random_state}{this parameter defines the seed of the random number generator, use when shuffling to mix observations.}
}
\value{
an instance of model
}
\description{
This function allows us to create a binary logistic regression model
}
\examples{
\dontrun{
 fit(formule, data)
 fit(formule, data, batch_size=10)
}
}
