% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_descent_batch.R
\name{grad_descent_batch}
\alias{grad_descent_batch}
\title{Gradient descent algorithm}
\usage{
grad_descent_batch(
  X,
  y,
  theta,
  leaning_rate = 0.1,
  max_iter = 100,
  tolerance = 1e-04
)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict.}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update.}

\item{max_iter}{is the number of iterations.}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}
}
\value{
this function returns the instance of model with all parameters
}
\description{
Gradient descent algorithm
}
\examples{
\dontrun{
  gradient(X,y,theta)
  gradient(X,y,theta, leaning_rate=0.1, max_iter=100, tolerance=1e-04)
}
}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
