% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{The gradient of the objective function}
\usage{
gradient(theta, X, y)
}
\arguments{
\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}
}
\value{
this function returns a gradient vector
}
\description{
The gradient function represents the partial derivative of the cost function with respect to each coefficient of the model
}
\examples{
\dontrun{
  gradient(theta, X, y)
}
}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
