% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradientElasticnet}
\alias{gradientElasticnet}
\title{Gradient for Elasticnet Loss Function}
\usage{
gradientElasticnet(theta, X, y, rho, C)
}
\arguments{
\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}

\item{rho}{hyper parameter which allows arbitration between RDIGE and LASSO}

\item{C}{parameter allowing to arbitrate between the penalty and the likelihood in the guidance of the modeling}
}
\value{
a vector
}
\description{
Gradient for Elasticnet Loss Function
}
\examples{
\dontrun{
  gradientElasticnet(theta, X, y,l1,l2)
}
}
