% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLoss.R
\name{logLoss}
\alias{logLoss}
\title{Logistic regression cost function}
\usage{
logLoss(theta, X, y)
}
\arguments{
\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}
}
\value{
this function returns a scalar corresponding to the cost for this vector of parameters \code{theta}
}
\description{
The cost function, or loss function is the function to be minimized(or maximized) by varying the decision variables.
}
\examples{
\dontrun{
  log_loss(theta, X, y)
}
}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
