% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLoss.R
\name{logLossElasticnet}
\alias{logLossElasticnet}
\title{Title}
\usage{
logLossElasticnet(theta, X, y, rho, C)
}
\arguments{
\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}

\item{rho}{hyper parameter which allows arbitration between RDIGE and LASSO}

\item{C}{parameter allowing to arbitrate between the penalty and the likelihood in the guidance of the modeling}
}
\value{
a scalar of the penalized cost function
}
\description{
Title
}
\examples{
\dontrun{
  logLossElasticnet(theta, X, y,l1,l2)
}
}
