% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_R2.R
\name{metric_R2}
\alias{metric_R2}
\title{coefficient of determination R2}
\usage{
metric_R2(y, ypred)
}
\arguments{
\item{y}{the true y}

\item{ypred}{the predict y}
}
\value{
accuracy or recognition rate
}
\description{
This function measures the performance of the model in terms of recognition rate.
}
\examples{
\dontrun{
   metric_R2(y, ypred)
}

}
\author{
"Saliou NDAO \href{mailto:salioundao21@gmail.com}{salioundao21@gmail.com}"
}
