% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics}
\alias{metrics}
\title{Metrics Function}
\usage{
metrics(pi, y, p)
}
\arguments{
\item{pi}{linear combination of predictor variables}

\item{y}{target variable to predict}

\item{p}{number of model parameters}
}
\value{
an S3 instance containing the values for the AIC (Akaike information criterion), the null deviance and the residual deviance
}
\description{
Metrics Function
}
\examples{
\dontrun{
  metrics(pi,y,p)
}
}
