% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Binary or probabilities prediction}
\usage{
predict(model, new_data, type_pred)
}
\arguments{
\item{model}{the model created with fit function logistic regression}

\item{new_data}{the new data to be classified}

\item{type_pred}{the type of prediction to perform (binary prediction or probability prediction)}
}
\value{
an instance containing the binary predictions, probability prediction, metrics.
}
\description{
Binary or probabilities prediction
}
\examples{
\dontrun{
 predict(model, new_data, type_pred)
}
}
