% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.modele.R
\name{print.modele}
\alias{print.modele}
\title{Customization function of the print method for the model instance}
\usage{
\method{print}{modele}(x, ...)
}
\arguments{
\item{x}{model instance returned by dgrglm.fit function}

\item{...}{other params}
}
\description{
Customization function of the print method for the model instance
}
