% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcharge_affichage_predict.R
\name{print.predict}
\alias{print.predict}
\title{Customization function of the print method for the predict instance}
\usage{
\method{print}{predict}(x, ...)
}
\arguments{
\item{x}{predict instance returned by dgrglm.predict function}

\item{...}{other argument}
}
\description{
Customization function of the print method for the predict instance
}
