% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage.quali.R
\name{recodage_X}
\alias{recodage_X}
\title{Re-coding Features}
\usage{
recodage_X(X)
}
\arguments{
\item{X}{data frame}
}
\value{
A encoding data frame
}
\description{
Re-coding Features
}
\examples{
recodage_X(iris)
}
\author{
Saliou NDAO
}
