% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(z)
}
\arguments{
\item{z}{linear combination of predictor variables}
}
\value{
a vector containing the probabilities of assignment to the positive or negative class
}
\description{
Sigmoid Function
}
\examples{
\dontrun{
  sigmoid(z)
}
}
