% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modele.R
\name{summary.modele}
\alias{summary.modele}
\title{Customization function of the summary method for the modele instance}
\usage{
\method{summary}{modele}(object, ...)
}
\arguments{
\item{object}{model instance returned by dgrglm.fit function}

\item{...}{other argument}
}
\description{
Customization function of the summary method for the modele instance
}
