% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcharge_affichage_predict.R
\name{summary.predict}
\alias{summary.predict}
\title{Customization function of the summary method for the predict instance}
\usage{
\method{summary}{predict}(object, ...)
}
\arguments{
\item{object}{predict instance returned by dgrglm.predict function}

\item{...}{other argument}
}
\description{
Customization function of the summary method for the predict instance
}
