% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.selection.R
\name{var.selection}
\alias{var.selection}
\title{Features Selection}
\usage{
var.selection(X, y, pval)
}
\arguments{
\item{X}{explicatives variables}

\item{y}{target variable}

\item{pval}{p-value}
}
\value{
an instance of selected features with associated statistics
}
\description{
Features Selection
}
\examples{
\dontrun{
  var.selection(X,y,pval)
}
}
