% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgsrow_batch_parallele.R
\name{dgsrow_batch_parallele}
\alias{dgsrow_batch_parallele}
\title{Batch DGSRow Distributed}
\usage{
dgsrow_batch_parallele(X, y, theta, ncores, leaning_rate, max_iter, tolerance)
}
\arguments{
\item{X}{is the matrix of our predictor variables with the bias column}

\item{y}{is a target variable to predict}

\item{theta}{is a vector containing the parameters or coefficient of the logistic to be estimated}

\item{ncores}{parameters representing the number of cores to be used for parallel execution}

\item{leaning_rate}{is the learning rate that controls the magnitude of the vector update}

\item{max_iter}{is the number of iterations}

\item{tolerance}{an additional parameter which specifies the minimum movement allowed for each iteration}
}
\value{
this function returns an instance containing:
\itemize{
\item final theta
\item history cost
\item iteration number
}
}
\description{
Batch DGSRow Distributed
}
\examples{
\dontrun{
 dgsrow_batch_parallele(X,y,theta)
 dgsrow_batch_parallele(X,y,theta,ncores=3)
}
}
