% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage.quali.R
\name{recodage.quali}
\alias{recodage.quali}
\title{Recoding target variable}
\usage{
recodage.quali(y)
}
\arguments{
\item{y}{qualitative variable to coded}
}
\value{
a dataframe where each modality represents a binary column
}
\description{
Recoding target variable
}
\examples{
recodage.quali(iris$Species)
}
\author{
Saliou NDAO
}
